chore: refactor the rate limit middleware to the top level lib.
